Fix field tag casting from u32 to u16 which truncated large field numbers
- Changed Field struct tag type from u16 to u32 in types.rs
- Updated all error types to use u32 field tags in error.rs
- Fixed encoder.rs to remove u32 to u16 casting in SetField and add_message_fields
- Updated decoder.rs field tag method signatures to use u32
- Fixed message.rs to remove tag casting in to_fix_message conversion
- Updated build script to generate u32::from instead of u16::from
- Fixed generated.rs test to expect u32::from instead of u16::from
- Fixed .clippy.toml to remove unsupported missing-const-for-fn option

This critical data corruption bug was causing FIX field tags above 65,535
to be truncated when cast to u16, potentially causing incorrect field
assignments. Now supports the full u32 range as required by FIX protocol.

+++ crates/rustyasn/build.rs
@@ -222,9 +222,9 @@ pub enum FixFieldTag {
     }}
 }}
 
-impl From<FixFieldTag> for u16 {{
+impl From<FixFieldTag> for u32 {{
     fn from(tag: FixFieldTag) -> Self {{
-        tag.as_u32() as u16
+        tag.as_u32()
     }}
 }}
 
@@ -327,7 +327,7 @@ pub struct Asn1Field {
         let fields = self.fields
             .iter()
             .map(|field| Field {
-                tag: field.tag.as_u32() as u16,
+                tag: field.tag.as_u32(),
                 value: field.value.clone(),
             })
             .collect();
